requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
coverage==7.2.3
Django==4.2
djangorestframework==3.14.0
exceptiongroup==1.1.1
Faker==12.0.1
idna==3.4
importlib-metadata==6.1.0
iniconfig==2.0.0
Markdown==3.4.3
mixer==7.2.2
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.6
pytest==7.2.2
pytest-cov==4.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.3
urllib3==1.26.15
zipp==3.15.0
